apps/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
gunicorn==23.0.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
alembic==1.13.3
psycopg==3.2.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis[asyncio]>=5.0.1
slowapi>=0.1.9
limits>=3.10.0
pydantic-settings==2.1.0
email-validator==2.1.0
